9-SiamFCpp/SiamFCpp/requirements.txt
tqdm
pylint
yacs
pillow
future
Cython==0.27.3
yapf==0.28.0
opencv-python==4.1.0.25
numpy==1.16.0
torch==1.4.0
torchvision==0.5.0
colorama==0.3.7
numba==0.45.0
matplotlib==3.0.0
wget==3.2
shapely==1.7.0
scipy==1.2.1
tensorboard
future==0.18.2
loguru==0.4.1
scikit-image
pycocotools
pandas
